+++ src/BabyKusto.Core/Evaluation/BuiltIns/BuiltInAggregates.cs
@@ -39,7 +39,7 @@ namespace BabyKusto.Core.Evaluation.BuiltIns
             aggregates.Add(
                 Aggregates.Sum,
                 new AggregateInfo(
-                    new AggregateOverloadInfo(new SumAggregateIntImpl(), ScalarTypes.Int, ScalarTypes.Int),
+                    new AggregateOverloadInfo(new SumAggregateIntImpl(), ScalarTypes.Long, ScalarTypes.Int),
                     new AggregateOverloadInfo(new SumAggregateLongImpl(), ScalarTypes.Long, ScalarTypes.Long),
                     new AggregateOverloadInfo(new SumAggregateDoubleImpl(), ScalarTypes.Real, ScalarTypes.Real)));
